refactoring `main.py`   + config.set_single_threaded_blas   + config.set_logging   + utils.multiprocess
